requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.9.0.post1
fastapi==0.115.7
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-anthropic==0.3.4
langchain-community==0.3.15
langchain-core==0.3.31
langchain-google-genai==2.0.5
langchain-openai==0.2.9
langchain-text-splitters==0.3.5
langsmith==0.3.1
marshmallow==3.26.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
openai==1.58.1
orjson==3.10.15
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
pymupdf==1.25.2
pyparsing==3.2.1
python-dotenv==1.0.1
pyyaml==6.0.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
sqlalchemy==2.0.37
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zstandard==0.23.0
